Translate To Morse Code and USART
added Translate to MorseCode(forgot to check in) with a task that will send the message out to a pin using two timers, which turn eachother on untill, causing the watch to turn on/off the buzzer.

also added a uart with bluetooth to input/output data, all it does now is echos the message received from the USART1. but malloc is not working sometime causing a hard fault i think i had a memory leak in sendMessage(so not everything in sendMessageQueue must be dynamically created since im freeing it)

Stack size increase, needed for the timers.