Port Application to STM32F101CBT6
Port Application to STM32F101CBT6, changes to pins were made UART stays the same, the EXTI is also the same. chaninged inthe GPIO inces the C port is not on this device.

configuration:
needed to change start up code and the project variables as well as the project target. the GPIO change was only needed since this MCU does not have those pins